Moved multi-class tests to set up testing file
